App Configuration Provider Move load (#46370)
* Move load to it's own file

* moving helper methods to utils

* reorginizing tests

* readding after merge

* copilot review comments
