+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/aio/__init__.py
@@ -3,8 +3,8 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # -------------------------------------------------------------------------
-
-from ._azureappconfigurationproviderasync import load, AzureAppConfigurationProvider
+from ._async_load import load
+from ._azureappconfigurationproviderasync import AzureAppConfigurationProvider
 
 __all__ = [
     "load",
